Moved get_addrlist to multi-os
