+++ src/os/mod.rs
@@ -10,3 +10,6 @@ pub mod linux;
 // FreeBSD Operating System support
 #[cfg(target_os = "freebsd")]
 pub mod freebsd;
+
+// Multi-operating System Support
+pub mod multi;
