+++ src/os/multi/mod.rs
@@ -1,0 +1,5 @@
+//! Multi-Operating System Support
+// This modules includes codes compatible over multiple operating systems
+
+// Standard C library
+pub mod libc;
